Fixed Snowcrash version number requirement
Blueman was throwing an exception on files created with the latest API Blueprint tools like Drafter because of the version number requirement. Removed this requirement and updated the code to support the AST format. Also updated the docs to make clear that Blueman only supports the AST format (which has been deprecated).